Moved automations API logic out of the endpoint definition
The initial version had a ton of persistence logic in the endpoint
itself; this pulls it out into a new
services/automations/automations-api.ts file.

This takes care of the @internal types (by switching to Typescript and
only exporting relevant types). Now the endpoint file is very simple
like most other endpoints in Ghost, and all the crazy fake automations
database stuff lives in the automations service directory.
